Add doc links and do not ask if only one profile exists

+++ pearl-config/install.sh
@@ -1,6 +1,8 @@
 post_install() {
     local profile_path=$(_detect_profile_path)
 
+    [[ -z "${profile_path}" ]] && { warn "No Firefox profile path found."; return; }
+
     link_to "$PEARL_PKGDIR"/user.js "${profile_path}"/user.js
     return 0
 }
@@ -12,6 +14,8 @@ post_update() {
 
 pre_remove() {
     local profile_path=$(_detect_profile_path)
+
+    [[ -z "${profile_path}" ]] && return
 
     rm -rf "${profile_path}"/user.js
     return 0
@@ -30,6 +34,10 @@ _detect_profile_path() {
     do
         profile_list+=("$profile")
     done
+
+    # If there is only one profile just do not ask
+    [[ "${#profile_list[@]}" == "1" ]] && { echo "${profile_list[0]}"; return; }
+    [[ "${#profile_list[@]}" == "0" ]] && return
 
     local profile_path=$(choose "Which Firefox profile?" "${profile_list[0]}" "${profile_list[@]}")
     echo "$profile_path"
